requirements.txt
streamlit
firebase-admin
requests
groq
python-dotenv